extract the logic into a util class for sync
